fix ash destroyer even more
i dug deep and figured out when and why vanilla mc added the setFire(1) line before doing dmg. its to make oneshotted mobs drop cooked meat instead of raw meat. was fixed in 1.4.4 quoting this bug report
https://bugs.mojang.com/browse/MC/issues/MC-1079

i dont see any reason not to use the full duration, as a failed hit will extinguish in either case.

the effect of this fix is making ash destroyer proc fully on first hit, and make it proc on iced dragonbone as well. idk if this should have a config toggle, but this is how it also behaved in 2.9.3 and before

special thanks to cdstk for bringing this to my attention